Added routes folder, fixed filenames and controllers
