+++ models/tagretolog.model.js
@@ -2,7 +2,6 @@ import mongoose from "mongoose";
 const Schema = mongoose.Schema;
 
 const TargetologSchema = new Schema({
-  _id: Schema.Types.ObjectId,
   name: { type: String, required: true },
   source: { type: String, required: true },
 });
