+++ package.json
@@ -14,7 +14,6 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.17.1",
-    "lodash": "^4.17.21",
     "mongoose": "^6.0.12",
     "nodemon": "^2.0.14"
   },
